AVG turnover rate
