+++ src/ClosedPositions/index.js
@@ -10,7 +10,6 @@ import './index.scss';
 
 const ClosedPositions = ({ summary, dispatch, sortConditions }) => {
   const renderIndividual = ({s}) => {
-    console.log(s.toJS())
     const profit = s.get('profit');
     return (
       <tr className='closedStockSummary__tr' key={`closedStockSummary__tr--${s.get('stockSymbol')}`}>
