+++ src/commons/getDaysDifference.js
@@ -3,8 +3,6 @@ const getDaysDifference = ({ day1, day2 }) => {
   d1.setHours(0,0,0,0)
   const d2 = new Date(day2);
   d2.setHours(0,0,0,0)
-  console.log(day2, day1)
-  console.log((d2 - d1)/1000/2600/24);
   return Math.floor((d2 - d1)/1000/2600/24);
 };
